Release updates

+++ app/ethWallet-geth.js
@@ -16,15 +16,11 @@ const envNet = process.env.ETHEREUM_NETWORK || 'mainnet'
 const envSubDomain = envNet === 'mainnet' ? 'ethwallet' : 'ethwallet-test'
 const gethDataDir = path.join(app.getPath('userData'), 'ethereum', envNet)
 
-const gethProcessKey = process.platform === 'win32'
-  ? 'geth.exe'
-  : 'geth'
-const ipcPath = process.platform === 'win32'
-  ? '\\\\.\\pipe\\geth.ipc'
-  : path.join(gethDataDir, 'geth.ipc')
-const pidPath = process.platform === 'win32'
-  ? '\\\\.\\pipe\\geth.pid'
-  : path.join(gethDataDir, 'geth.pid')
+const isWindows = process.platform === 'win32'
+const gethProcessKey = isWindows ? 'geth.exe' : 'geth'
+
+const ipcPath = isWindows ? '\\\\.\\pipe\\geth.ipc' : path.join(gethDataDir, 'geth.ipc')
+const pidPath = isWindows ? '\\\\.\\pipe\\geth.pid' : path.join(gethDataDir, 'geth.pid')
 const gethProcessPath = path.join(getExtensionsPath('bin'), gethProcessKey)
 
 const configurePeers = async (dataDir) => {
@@ -95,7 +91,8 @@ const spawnGeth = async () => {
     '--ipcpath',
     ipcPath,
     '--maxpeers',
-    '10'
+    '10',
+    '--nousb'
   ]
 
   if (envNet === 'ropsten') {
@@ -156,6 +153,7 @@ const writeGethPid = async (pid) => {
   gethProcessId = pid
 
   try {
+    await fs.ensureDir(gethDataDir)
     await fs.writeFile(pidPath, gethProcessId)
   } catch (ex) {
     console.error('Could not write geth.pid')
@@ -177,10 +175,14 @@ const cleanupGeth = (processId) => {
       gethProcessId = null
     }
 
-    try {
-      fs.unlinkSync(pidPath)
-    } catch (ex) {
-      console.error('Could not delete geth.pid')
+    // Named pipes on Windows will get deleted
+    // automatically once no processes are using them.
+    if (!isWindows) {
+      try {
+        fs.unlinkSync(pidPath)
+      } catch (ex) {
+        console.error('Could not delete geth.pid')
+      }
     }
     console.warn('GETH: cleanup done')
   }
